+++ custom_components/plugwise_usb/coordinator.py
@@ -87,7 +87,9 @@ class PlugwiseUSBDataUpdateCoordinator(DataUpdateCoordinator):
             and self.node.initialized
             and not states[NodeFeature.AVAILABLE].state
         ):
-            raise UpdateFailed("Device is not responding")
+            raise UpdateFailed(
+                f"Device '{self.node.node_info.mac}' is (temporarily) not available"
+            )
 
         return states